Feat/integrate inventory

+++ app/src/main/java/bangkit/kiki/foodwisemobile/data/model/UserStatisticsModel.kt
@@ -6,9 +6,9 @@ data class UserInventoryResponse(
     @SerializedName("consumed_count") val consumedCount: Int,
     @SerializedName("in_stock_count") val inStockCount: Int,
     @SerializedName("expired_count") val expiredCount: Int,
-    @SerializedName("current_progress") val currentProgress: Int, // current progress in percentage
+    @SerializedName("current_progress") val currentProgress: Float, // current progress in percentage
     @SerializedName("remaining_days") val remainingDays: Int, // remaining days in the current month cycle
-    @SerializedName("history_progress") val historyProgress: Int? // history progress in percentage, nullable
+    @SerializedName("history_progress") val historyProgress: Float? // history progress in percentage, nullable
 )
 
 data class ExpiringFood(
